refactor(sdk): reorganize utility functions into public/internal
Reorganized utility functions in the SDK to clearly separate public and
internal APIs:
- Created a new `internal` directory for utility functions not meant for
  public use
- Moved `array.ts`, `bytes.ts`, and `state.ts` to the internal directory
- Updated import paths across the codebase
- Updated `utils/README.md` to document all public utility functions

This change improves maintainability and provides clearer API boundaries
for SDK users.
