+++ packages/sdk/src/utils/internal/bytes.ts
