+++ packages/sdk/src/utils/internal/state.ts
